set figure size to integer

+++ plotting.py
@@ -35,7 +35,7 @@ def circular_depgraph(g,
 	#here we do black magic to get proper output size (controls vertex spacing) and scaling
 	vertex_number = g.num_vertices()
 	view_zoom = (vertex_number*36.0485)**(-10.068/vertex_number)+0.017037
-	output_size = vertex_number*5.9+400
+	output_size = int(vertex_number*5.9+400)
 	dpi=300
 	if output_size >= 18000:
 		print("WARNING: You are exceding the maximal printable size - 150cm in one dimension at 300dpi")
